+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 +++++
 
 - D3 7.4.0 support
+- Remove issue autowatch (this feature comes with Redmine 5)
 
 3.0.5.2
 +++++++